Refactor common queue methods and classes
Pull out a QueueCommon class that the various threads can inherit
from and unify the Command class in the same file. Also fixes a
bunch of instance variable initialization in the main thread
classes. Other files have not been fixed, but since almost
everything else is only used once it shouldn't be an issue.